Add missing optional field to NFTokenBurn
When burning a token with lsfBurnable enabled as the original minter, you need to specify the current `Owner` of the NFToken. Otherwise `rippled` will default to searching the normal `Account` for the NFToken and will return `tecNO_ENTRY` since someone else has the NFT.

I haven't worked with this repo before, so I may be missing tests/additional logic which is required to make this change work. Please let me know if that is the case!

Opened in response to this issue: https://github.com/XRPLF/xrpl.js/issues/1947