Remove redundant array copy
Calling UBA.toByteArray() will copy the bytes from the UBA inot a newly initialized byte array. This code was doing this twice, just to check the legnth. In addition, the length of the UBA was being checked, whereas the length of the byte[] should have been checked instead. Generally, these two lengths should be the same, but for correctness the byte[] length is preferred.